Speed up tests
