+++ tests/Integration/Subscriptions/Broadcast/BroadcastDBTest.php
@@ -1,15 +1,14 @@
 <?php
 
-namespace Tests\Unit\Subscriptions\Broadcast;
+namespace Tests\Integration\Subscriptions\Broadcast;
 
 use Tests\DBTestCase;
-use Mockery\MockInterface;
 use Tests\Utils\Models\Task;
 use Nuwave\Lighthouse\Execution\Utils\Subscription;
 use Nuwave\Lighthouse\Subscriptions\SubscriptionBroadcaster;
 use Nuwave\Lighthouse\Subscriptions\SubscriptionServiceProvider;
 
-class BroadcastTest extends DBTestCase
+class BroadcastDBTest extends DBTestCase
 {
     protected $schema = '
     type Task {
@@ -45,24 +44,12 @@ class BroadcastTest extends DBTestCase
         factory(Task::class, 1)->create();
     }
 
-    /**
-     * @return MockInterface
-     */
-    public function withMockedBroadcasts(): MockInterface
-    {
-        $broadcast = \Mockery::mock(SubscriptionBroadcaster::class);
-        $this->app->instance(SubscriptionBroadcaster::class, $broadcast);
-
-        return $broadcast;
-    }
-
     public function testBroadcastsFromPhp(): void
     {
-        // Assert
-        $broadcast = $this->withMockedBroadcasts();
-        $broadcast->shouldReceive('broadcast')->once();
+        $this->mock(SubscriptionBroadcaster::class)
+            ->shouldReceive('broadcast')
+            ->once();
 
-        // Subscribe to event
         $this->postGraphQL([
             'query' => '
                 subscription UserUpdated {
@@ -73,17 +60,15 @@ class BroadcastTest extends DBTestCase
             ',
         ]);
 
-        // Broadcast
         Subscription::broadcast('taskUpdated', []);
     }
 
     public function testBroadcastsFromSchema(): void
     {
-        // Assert
-        $broadcast = $this->withMockedBroadcasts();
-        $broadcast->shouldReceive('broadcast')->once();
+        $this->mock(SubscriptionBroadcaster::class)
+            ->shouldReceive('broadcast')
+            ->once();
 
-        // Subscribe to event
         $this->postGraphQL([
             'query' => '
                 subscription TaskUpdated {
@@ -94,7 +79,6 @@ class BroadcastTest extends DBTestCase
             ',
         ]);
 
-        // Broadcast
         $this->postGraphQL([
             'query' => '
                 mutation {
