+++ tests/Unit/Schema/Directives/PaginateDirectiveTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Unit\Schema\Directives;
 
+use GraphQL\Error\Error;
 use Tests\TestCase;
 use GraphQL\Type\Definition\FieldArgument;
 use GraphQL\Type\Definition\FieldDefinition;
@@ -147,5 +148,119 @@ class PaginateDirectiveTest extends TestCase
             FieldArgument::class,
             $queryType->getField('defaultPaginated')->getArg('first')
         );
+    }
+
+    public function testIsLimitedByMaxCountFromDirective(): void
+    {
+        config(['lighthouse.paginate_max_count' => 5]);
+
+        $this->schema = '
+        type User {
+            id: ID!
+            name: String!
+        }
+        
+        type Query {
+            users1: [User!]! @paginate(maxCount: 6)
+            users2: [User!]! @paginate(maxCount: 10)
+        }
+        ';
+
+        $result = $this->graphQL('
+        {
+            users1(first: 10) {
+                data {
+                    id
+                    name
+                }
+            }
+        }
+        ');
+
+        $this->assertSame(
+            'Maximum number of 6 requested items exceeded. Fetch smaller chunks.',
+            $result->jsonGet('errors.0.message')
+        );
+    }
+
+    public function testIsLimitedToMaxCountFromConfig(): void
+    {
+        config(['lighthouse.paginate_max_count' => 5]);
+
+        $this->schema = '
+        type User {
+            id: ID!
+            name: String!
+        }
+        
+        type Query {
+            users1: [User!]! @paginate
+            users2: [User!]! @paginate(type: "relay")
+        }
+        ';
+
+        $resultFromDefaultPagination = $this->graphQL('
+        {
+            users1(first: 10) {
+                data {
+                    id
+                    name
+                }
+            }
+        }
+        ');
+
+        $this->assertSame(
+            'Maximum number of 5 requested items exceeded. Fetch smaller chunks.',
+            $resultFromDefaultPagination->jsonGet('errors.0.message')
+        );
+
+        $resultFromRelayPagination = $this->graphQL('
+        {
+            users2(first: 10) {
+                edges {
+                    node {
+                        id
+                        name
+                    }
+                }
+            }
+        }
+        ');
+
+        $this->assertSame(
+            'Maximum number of 5 requested items exceeded. Fetch smaller chunks.',
+            $resultFromRelayPagination->jsonGet('errors.0.message')
+        );
+    }
+
+    public function testThrowsWhenPaginationWithCountZeroIsRequested(): void
+    {
+        $this->schema = '
+        type User {
+            id: ID!
+            name: String!
+        }
+        
+        type Query {
+            users: [User!] @paginate
+        }
+        ';
+
+        $this->graphQL('
+        {
+            users(first: 0) {
+                data {
+                    id
+                }
+            }
+        }
+        ')
+        ->assertJson([
+            'data' => [
+                'users' => null,
+            ],
+        ])
+        ->assertErrorCategory(Error::CATEGORY_GRAPHQL);
     }
 }
