+++ tests/Unit/Schema/Directives/SpreadDirectiveTest.php
@@ -2,78 +2,10 @@
 
 namespace Tests\Unit\Schema\Directives;
 
-use Tests\DBTestCase;
-use Tests\Utils\Models\User;
+use Tests\TestCase;
 
-class SpreadDirectiveTest extends DBTestCase
+class SpreadDirectiveTest extends TestCase
 {
-    public function testSpreadsTheInputIntoTheQuery(): void
-    {
-        factory(User::class, 2)->create();
-
-        $this->schema = '
-        type Query {
-            user(input: UserInput @spread): User @first
-        }
-
-        type User {
-            id: ID
-        }
-
-        input UserInput {
-            id: ID @eq
-        }
-        ';
-
-        $this->graphQL('
-        {
-            user(input: {
-                id: 2
-            }) {
-                id
-            }
-        }
-        ')->assertJson([
-            'data' => [
-                'user' => [
-                    'id' => 2,
-                ],
-            ],
-        ]);
-    }
-
-    public function testIgnoresSpreadedInputIfNotGiven(): void
-    {
-        factory(User::class)->create();
-
-        $this->schema = '
-        type Query {
-            user(input: UserInput @spread): User @first
-        }
-
-        type User {
-            id: ID
-        }
-
-        input UserInput {
-            id: ID @eq
-        }
-        ';
-
-        $this->graphQL('
-        {
-            user {
-                id
-            }
-        }
-        ')->assertJson([
-            'data' => [
-                'user' => [
-                    'id' => 1,
-                ],
-            ],
-        ]);
-    }
 
     public function testNestedSpread(): void
     {
