+++ tests/Unit/Subscriptions/StorageManagerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Integration\Subscriptions;
+namespace Tests\Unit\Subscriptions;
 
 use Tests\TestCase;
 use GraphQL\Utils\AST;
